move showRunning to WorkHandler.onProgress
